requirements.txt
instagrapi==2.2.1
openai==1.102.0
pexels_api==1.0.1
Pillow==9.5.0
Pillow==11.3.0
python-dotenv==1.1.1
Requests==2.32.5
schedule==1.2.2
streamlit==1.48.0
